add chapter headings and developer guide to Sphinx docs (#465)
### Description
This is a major update to the Sphinx documentation that rewrites the
introduction to be more consistent with `README.md`, adds chapter
headings for the User Guide and the Developer Guide, and moves the
developer-oriented content from the GitHub wiki to the Sphinx
documentation.

### Related issues
N/A

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues see (see above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [ ] I have added tests for any new physics that this PR adds to the
code.
- [x] I have tested this PR on my local computer and all tests pass.
- [x] I have manually triggered the GPU tests with the magic comment
`/azp run`.
- [x] I have requested a reviewer for this PR.