add support for writing openPMD/ADIOS2 plotfiles (#466)
### Description
This adds support for writing ADIOS2 format plotfiles using openPMD-api
(https://github.com/openPMD/openPMD-api). This is the file format used
by WarpX for its large-scale runs, and it is the fastest option for
large-scale I/O (~Terabytes or larger).

This adds support for cell-centered variables only. (However, it does
include cell-centered averages of the face-centered variables, and all
of the other derived variables.)

openPMD currently does not support ghost cells, so we exclude them when
writing the plotfile in openPMD/ADIOS2 format.

There is no support for checkpoints or restarts with this format (nor is
there in WarpX, or any other code that I'm aware of).

### Related issues
N/A

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues see (see above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [x] I have added tests for any new physics that this PR adds to the
code.
- [x] I have tested this PR on my local computer and all tests pass.
- [x] I have manually triggered the GPU tests with the magic comment
`/azp run`.
- [x] I have requested a reviewer for this PR.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>