Fallback to zero-gradient extrapolation when NSCBC fails (#472)
### Description
This implemented a fallback to zero-gradient extrapolation when the
NSCBC outflow boundary condition produces ghost cells with negative
density or pressure.

This is necessary to prevent the simulation from crashing when NSCBC
outflow is used for the cloud acceleration problem. There does not
appear to be anything better that can be done when this happens
(Zingale, personal communication). This only is needed for an extremely
small fraction of ghost cells.

### Related issues
N/A

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues see (see above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [x] I have added tests for any new physics that this PR adds to the
code.
- [x] I have tested this PR on my local computer and all tests pass.
- [x] I have manually triggered the GPU tests with the magic comment
`/azp run`.
- [x] I have requested a reviewer for this PR.