add tracer particles (#490)
### Description
This adds tracer particles. Particles are created one-per-cell in the
initial conditions, then they advect with the fluid using the
face-centered velocities returned from the Riemann solver. Particles are
included in plotfiles and checkpoints. *(Particles will be added to
openPMD outputs in a subsequent PR.)*

The method adopted here for particle advection with face-centered
velocities is the ["marker-and-cell method" (Harlow & Welch
1965)](https://ui.adsabs.harvard.edu/abs/1965PhFl....8.2182H/abstract).

Tracer particles are turned on for the `HydroBlast3D` and
`SphericalCollapse` problems, so tracer particle functionality is
included in the test suite.

*Significant code infrastructure changes:*
* This removes the old AMR level subcycling code taken from Chombo and
replaces it with the cleaner version from the AMReX code examples.
However, it should behave identically as before. This is necessary for
particle redistribution to work with subcycling.
* This adds the face-centered velocities to `state_new_fc_` and
`state_old_fc_`. Some changes to the machinery for face-centered
variables was necessary to get this to work, such as implementing AMR
interpolation for face-centered variables using the `amrex::FaceLinear`
interpolator (*caution: this interpolator does NOT preserve the
divergence of the field!*). This is necessary because ghost faces are
needed in order to handle particles that cross coarse-fine boundaries
during a timestep. (In the future, this machinery can be modified so
that it creates alias MultiFabs for separate face-centered components
and then uses a different AMR interpolator for each component.)
* In order to avoid modifying the initial conditions for every problem
when tracer particles are enabled, this separates the initial conditions
set-up function into `setInitialConditionsOnGrid` (cell-centered
variables) and `setInitialConditionsOnGridFaceVars` (face-centered
variables). This only affects the `FCQuantities` test problem.

### Related issues
Closes https://github.com/quokka-astro/quokka/issues/489 and
https://github.com/quokka-astro/quokka/issues/498.

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues see (see above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [x] I have added tests for any new physics that this PR adds to the
code.
- [x] I have tested this PR on my local computer and all tests pass.
- [x] I have manually triggered the GPU tests with the magic comment
`/azp run`.
- [x] I have requested a reviewer for this PR.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>