fix copies to rank zero for correctness tests (#531)
### Description
This fixes parallel copies of particles and/or MultiFabs to rank zero so
that they actually work. In this case, constructing a
`amrex::DistributionMapping` must be done by explicitly setting it equal
to `{0}`.

This is used when testing the correctness of the results for the
`HydroRichtmyerMeshkov` and `BinaryOrbitCIC` tests.

Thanks to Andrew Myers for the fix.

### Related issues
N/A

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues see (see above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [ ] I have added tests for any new physics that this PR adds to the
code.
- [x] I have tested this PR on my local computer and all tests pass.
- [x] I have manually triggered the GPU tests with the magic comment
`/azp run`.
- [x] I have requested a reviewer for this PR.