Use fully implicity scheme for flux update
Before this PR, the gas momentum is delayed when updating radiation flux using backward Euler scheme. Now, a set of 3N + 3 linear equations is solved for the updated v and F_{r, i}. This is fully tested on all test problems with success. However, this implementation brings in a lot more computation but does not significantly change the result of any of the current tests, therefore I keep it disabled by default. I have enabled it in the shock_multigroup test.