Define K_S for isothermal gamma (#573)
### Description
Currently, `K_S` is undefined in the `computeFlatenningCoefficients`
function for the case of isothermal gamma, because the function to
compute sound speed within `EOS.hpp` that it calls only works if gamma
!= 1.

### Related issues
Fixes #553 for now. In the long run, this wont be needed because we will
move all EOS calculations (including for gamma=1) to EOS.hpp. But for
now, we should apply this patch so that the gamma=1 case is correctly
handled while computing flattening coefficients.

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues see (see above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [ ] I have added tests for any new physics that this PR adds to the
code.
- [x] I have tested this PR on my local computer and all tests pass.
- [x] I have manually triggered the GPU tests with the magic comment
`/azp run`.
- [x] I have requested a reviewer for this PR.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>