Update test parameters to produce all the figures in the IMEX paper (#600)
### Description

Under the reproducible research philosophy which I believe in, I would
like the reader to be able to reproduce all the figures and data from a
paper with little effort. Therefore, I updated the tests used in the
IMEX paper, setting some of the physical parameters as runtime
parameters, so that the reader can reproduce the exact result in the
paper by simply changing the config.in files, while the default
parameters are kept moderate so that the automatic tests can run in a
reasonable time.

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues see (see above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [x] I have added tests for any new physics that this PR adds to the
code.
- [x] I have tested this PR on my local computer and all tests pass.
- [x] I have manually triggered the GPU tests with the magic comment
`/azp run`.
- [x] I have requested a reviewer for this PR.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>