run test suite with Intel oneAPI compiler on CPU (#618)
### Description
This adds a GitHub action that compiles Quokka and runs the test suite
with the Intel oneAPI compiler on CPU. This is needed in order to use a
compute allocation on Stampede3.

Note that this compiler is LLVM-based and is *not* the old Intel `icc`
compiler (which cannot compile Quokka).

### Related issues
N/A

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues see (see above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [ ] I have added tests for any new physics that this PR adds to the
code.
- [ ] I have tested this PR on my local computer and all tests pass.
- [ ] I have manually triggered the GPU tests with the magic comment
`/azp run`.
- [ ] I have requested a reviewer for this PR.