change initial condition of RadForce (#627)
### Description

The RadForce test can be improved to be more realistic and less `ad
hoc`. Instead of initializing with the exact solution and letting the
simulation maintain a static state, we start the simulation with a
uniform density of rho0 and zero velocity. The left boundary condition
is kept the same as before, but the right boundary is set to be
foextrap. The end result is exactly the same as before: the radiation
blows away the gas and the system comes to the static state given by the
differential equation.

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [ ] I have added a link to any related issues see (see above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [x] I have added tests for any new physics that this PR adds to the
code.
- [x] I have tested this PR on my local computer and all tests pass.
- [ ] I have manually triggered the GPU tests with the magic comment
`/azp run`.
- [x] I have requested a reviewer for this PR.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>