Fix GPU problems in pulse_MG.cpp  (#629)
### Description

I discovered some GPU-related issues in radhydro_pulse_MG.cpp and fixed
it. In the previous PR, I turned if off for 3D problems because it is
too expensive. As a result, it didn't get to compile and run on nvcc or
ROCm. I fixed some first-capture and undefined on device issues and
turned this test (and radhydro_pulse_MG_int.cpp) on. I set
`max_timesteps=10` in these tests in order to let them finish fast. We
are not comparing with exact solutions any way (because there isn't
one), so it's okay to run only a small number of time steps and show
that the architecture is working. The accuracy of the multigroup
implementation is tested by hand, and I will include it in a planned
nightly unit test.

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues see (see above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [ ] I have added tests for any new physics that this PR adds to the
code.
- [x] I have tested this PR on my local computer and all tests pass.
- [x] I have manually triggered the GPU tests with the magic comment
`/azp run`.
- [x] I have requested a reviewer for this PR.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>