Clean up code for single-group radiation  (#684)
### Description

This PR updates the code structure for single-group radiation so that
functions like `ComputePlanckOpacity` return a `double`. This change
enforces the use of the correct opacity models (i.e.,
`piecewise_constant_opacity` or `PPL_opacity_fixed_slope_spectrum`, or
`PPL_opacity_full_spectrum`) when dealing with multiple radiation
groups. The previous code provides an option to naively extend the
single-group formulation in the IMEX paper to the multigroup case, i.e.
replacing $\kappa_P$ and $\kappa_F$ with $\kappa_g$. However, that
method does not correctly describe multigroup radiation transport and
should be removed, because it misses the important delta term in the
flux equation. See Equation 16 of the multigroup paper and the
discussion thereafter.

In summary, here's how you should define gas opacity in a problem:
- Single-group opacity definition:
- Define `ComputePlanckOpacity` and optionally `ComputeFluxMeanOpacity`
and `ComputeEnergyMeanOpacity`. By default, the latter two are set equal
to the first.
- Multigroup opacity definition:
- Choose an opacity model by setting `opacity_model` in
`RadSystem_Traits` to one of the following:
`piecewise_constant_opacity`, `PPL_opacity_fixed_slope_spectrum`, or
`PPL_opacity_full_spectrum`
- Define `DefineOpacityExponentsAndLowerValues`. See the example in
RadhydroShockMultigroup.

Bug fix:
- fixed bug related to multigroup opacity in `ComputeCellOpticalDepth`.
This function is only used when you turn on wave speed correction by
hand and is not turned on by default. We can't have a regression test
for this because it is controlled by `use_wavespeed_correction` inside
`radiation_system.hpp` and we don't have a way to turn it on for a
specific problem. I tested it on my computer and it worked as expected.

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues see (see above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [x] I have added tests for any new physics that this PR adds to the
code.
- [x] I have tested this PR on my local computer and all tests pass.
- [x] I have manually triggered the GPU tests with the magic comment
`/azp run`.
- [x] I have requested a reviewer for this PR.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Ben Wibking <ben@wibking.com>