Particle creator (#896)
### Description

The `createCICParticles` method in `PhysicsParticles.hpp` provides a
framework for particle creation on the fly from fluid cells. We run
`createCICParticles` in the end of each hydro timestep. It uses two key
components:

1. `CICParticleChecker`: A functor that determines whether a particle
should be created in a given cell based on:
- State-based conditions (can be customized to check fluid properties)
   - Time conditions 
   - Spatial conditions (grid spacing)

2. `CICParticleCreator`: A functor that handles the actual particle
creation and initialization, including:
   - Setting particle position (e.g. cell-centered)
   - Assigning particle properties (mass, velocities)
   - Conserving mass and momentum between fluid and particles
   - Updating the fluid state after particle creation

This separation of checker and creator allows for easy customization of
both when/where particles are created and how they are initialized.
Currently, `CICParticleChecker` is set to return false, meaning no
particles are created. `CICParticleCreator` is left empty. In
`gravity_3d.cpp`, I showed an example of how to override the default
checker and creator.

This implementation is validated by the `gravity_3d` test on CPU, GPU,
and MPI, which demonstrates:
1. correct orbital dynamics of the CIC particles under self-gravity.
2. gravitational interaction between fluid, CIC particles, and
CICRadParticles.
3. energy conservation of radiation energy in the presence of
RadParticles.
4. creation of CICParticles from fluid cells on the fly.

#### Future

1. Particle creation with access to neighbor cells can be supported by
adding ghost cells to the multi-fab.
2. Particle-fluid interaction can be implemented similarly with very
little extra work.
3. More work needs to be done to add non-gravitational particle-particle
interaction. `NeighborParticleContainer` does not have a AMR
implementation, as far as I know.

### Related issues
None.

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues (if applicable; see
above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [x] I have added tests for any new physics that this PR adds to the
code.
- [ ] *(For quokka-astro org members)* I have manually triggered the GPU
tests with the magic comment `/azp run`.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>