Remove all unused const variables (#941)
### Description

This PR removes all unused const variables so that we can successfully
build the code with `-DWARNINGS_AS_ERRORS=ON` on moth. This helps to
reproduce the warning github action locally.

When I compile Quokka on moth with warnings_as_errors on, I got millions
of `-Wunused-const-variable` errors. I don't know why we don't see that
kind of warnings in the github action, but in order to avoid that on
moth, I have to remove all unused const variables.

### Related issues
Are there any GitHub issues that are fixed by this pull request? Add a
link to them here.

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues (if applicable; see
above).
- [ ] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [ ] I have added tests for any new physics that this PR adds to the
code.
- [ ] *(For quokka-astro org members)* I have manually triggered the GPU
tests with the magic comment `/azp run`.