Add Stellar Evolution attribute and SNDeposition (#938)
### Description

This PR adds a new attribute `evolutionStage` to the particle data and a
new function `SNDeposition` to deposit mass, momentum, and energy from
particles to the grid.

Changes:

1. Add `Test` particle type for testing all particle features. 

2. Define an integer component `evolutionStage` to Test_particles. It is
used to track the evolution stage of the particle. The evolution stage
is defined as an enum class as listed below. For particle types that
support stellar evolution (`evolutionStageIndex_ >= 0`), each star is
created either as a low mass star or as a SN progenitor. When a SN
progenitor reaches the end of its life, it executes `SNDeposition()` and
evolves into a SN remnant.

```cpp
enum class StellarEvolutionStage {
	LowMassStar,  // Low mass star stage
	SNProgenitor,  // Supernova progenitor stage
	SNRemnant     // Supernova remnant stage
};
```

3. Defined `SNDeposition` to deposit mass, momentum, and energy from
particles to the grid. At each time step, we find the particle types
that support stellar evolution, and check if each particle is a SN
progenitor and if it has reached the end of its life. If so, we run
`SNDeposition()` and update the particle's evolution stage to SNRemnant.
This also works as a framework for other interact-with-mesh operations.
Unlike `interp.ParticleToMesh` which uses linear interpolation to
deposit N particle components to N mesh components on a one-to-one
basis, `SNDeposition` calculates the exact position of the particle in
the mesh and deposit particle properties to arbitrary components of
neighboring cells, using any interpolation kernel. Note that the maximum
number of cells allowed to use in SNDeposition (the `stencil_width`) is
`nghost_cc_ = 4`.

The new code is fully tested via the gravity_3d test on CPU, GPU, and
MPI, where we check the creation, stellar evolution, and destruction of
the particles, as well as the total mass in SN ejecta.

### Checklist
_Before this pull request can be reviewed, all of these tasks should be
completed. Denote completed tasks with an `x` inside the square brackets
`[ ]` in the Markdown source below:_
- [x] I have added a description (see above).
- [x] I have added a link to any related issues (if applicable; see
above).
- [x] I have read the [Contributing
Guide](https://github.com/quokka-astro/quokka/blob/development/CONTRIBUTING.md).
- [x] I have added tests for any new physics that this PR adds to the
code.
- [ ] *(For quokka-astro org members)* I have manually triggered the GPU
tests with the magic comment `/azp run`.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>